backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
Django==6.0.3
django-cors-headers==4.9.0
dj-database-url==2.1.0
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
django-rest-passwordreset==1.5.0
google-auth==2.38.0
feedparser==6.0.12
idna==3.11
psycopg2-binary==2.9.11
PyJWT==2.12.1
requests==2.32.5
sgmllib3k==1.0.0
sqlparse==0.5.5
urllib3==2.6.3
python-dotenv==1.2.2
